validate with session

+++ backend/src/routes/user.ts
@@ -1,9 +1,12 @@
-import { Hono, Context } from 'hono'
-import bcrypt from 'bcryptjs'
-import { Env } from '../index'
-import { db } from '../repository'
-import { Query } from '../types'
-import { deleteImagesFromR2 } from '../r2'
+import bcrypt from "bcryptjs";
+import { Hono, Context } from "hono";
+
+import { setSessionUser, validateSession, clearSession } from "../../lib/auth";
+import { Env } from "../index";
+import { deleteImagesFromR2 } from "../r2";
+
+import { Query } from "../types";
+import { db } from "../repository";
 
 const userRoutes = new Hono<{ Bindings: Env }>();
 
@@ -14,6 +17,7 @@ const handleDemoEmail = (c: Context<{ Bindings: Env }>, email: string) => {
   return false;
 };
 
+
 // userRoutes.post('/signup', async (c) => {
 //   const { email, password, username }: User = await c.req.json()
 //   const hashedPassword = await bcrypt.hash(password, 10)
@@ -22,53 +26,85 @@ const handleDemoEmail = (c: Context<{ Bindings: Env }>, email: string) => {
 //   return c.json({ message: 'User created' })
 // })
 
-userRoutes.post('/signout', async (c) => {
-  const { email }: Query = await c.req.json()
-  if (handleDemoEmail(c, email)) {
-    return c.json({ message: 'Demo user cannot be deleted' }, 400);
+
+userRoutes.post("/signout", async (c) => {
+  const { email }: Query = await c.req.json();
+  if (!(await validateSession(c, email))) {
+    return c.json({ message: "Unauthorized" }, 401);
   }
-  const database = db(c.env)
-  await database.user.delete({ email })
+
+  if (handleDemoEmail(c, email)) {
+    return c.json({ message: "Demo user cannot be deleted" }, 400);
+  }
+  const database = db(c.env);
+  await database.user.delete({ email });
   await database.clothes.deleteByUser({ userID: email });
   await database.wearHistory.deleteByUser({ email });
   await deleteImagesFromR2(c.env, email);
-  return c.json({ message: 'User and related data deleted' }, 200)
-})
+  // delete session
+  await clearSession(c);
+  return c.json({ message: "User and related data deleted" }, 200);
+});
 
-userRoutes.post('/login', async (c) => {
+
+userRoutes.post("/login", async (c) => {
   const { email, password }: Query = await c.req.json();
   const database = db(c.env);
   const user = await database.user.findUnique({ email: email });
 
-  if (user && user.password && password && await bcrypt.compare(password, user.password)) {
+  if (
+    user &&
+    user.password &&
+    password &&
+    (await bcrypt.compare(password, user.password))
+  ) {
+    // save user in session
+    await setSessionUser(c, email);
     return c.json({ username: user.username }, 200);
   }
 
-  return c.json({ message: 'Invalid credentials' }, 401);
+  return c.json({ message: "Invalid credentials" }, 401);
 });
 
-userRoutes.post('/logout', (c) => {
-  return c.json({ message: 'Logout successful' }, 200);
-})
 
-userRoutes.post('/changePassword', async (c) => {
-  const { email, password, newPassword }: { email: string; password: string; newPassword: string } = await c.req.json();
+userRoutes.post("/logout", (c) => {
+  clearSession(c);
+  return c.json({ message: "Logout successful" }, 200);
+});
+
+
+userRoutes.post("/changePassword", async (c) => {
+  const {
+    email,
+    password,
+    newPassword,
+  }: { email: string; password: string; newPassword: string } =
+    await c.req.json();
+
+  if (!(await validateSession(c, email))) {
+    return c.json({ message: "Unauthorized" }, 401);
+  }
+
   if (handleDemoEmail(c, email)) {
-    return c.json({ message: 'Demo user cannot change password' }, 400);
+    return c.json({ message: "Demo user cannot change password" }, 400);
   }
   const database = db(c.env);
   const user = await database.user.findUnique({ email });
 
-  if (user && user.password && await bcrypt.compare(password, user.password)) {
+  if (
+    user &&
+    user.password &&
+    (await bcrypt.compare(password, user.password))
+  ) {
     if (password === newPassword) {
-      return c.json({ message: 'New password must be different' }, 400);
+      return c.json({ message: "New password must be different" }, 400);
     }
     const hashedNewPassword = await bcrypt.hash(newPassword, 10);
     await database.user.updatePassword({ email, password: hashedNewPassword });
-    return c.json({ message: 'Password changed' }, 200);
+    return c.json({ message: "Password changed" }, 200);
   }
 
-  return c.json({ message: 'Invalid credentials' }, 401);
+  return c.json({ message: "Invalid credentials" }, 401);
 });
 
 export default userRoutes;
